test(health): moves and renames healthcheck.spec.ts to get.test.ts
